+++ src/components/suggest/suggest_item.js
@@ -31,9 +31,16 @@ export const EuiSuggestItem = ({
   type,
   labelDisplay,
   description,
+  onClick,
   ...rest
 }) => {
-  const classes = classNames('euiSuggestItem', className);
+  const classes = classNames(
+    'euiSuggestItem',
+    {
+      'euiSuggestItem-isClickable': onClick,
+    },
+    className
+  );
 
   let colorClass = '';
 
@@ -51,14 +58,19 @@ export const EuiSuggestItem = ({
     }
   }
 
+  let OuterElement = 'div';
+  if (onClick) {
+    OuterElement = 'button';
+  }
+
   return (
-    <div className={classes} {...rest}>
+    <OuterElement onClick={onClick} className={classes} {...rest}>
       <span className={`euiSuggestItem__type ${colorClass}`}>
         <EuiIcon type={type.iconType} />
       </span>
       <span className={labelDisplayClass}>{label}</span>
       <span className="euiSuggestItem__description">{description}</span>
-    </div>
+    </OuterElement>
   );
 };
 
@@ -84,6 +96,10 @@ EuiSuggestItem.propTypes = {
    * Label display is 'fixed' by default. Label will increase its width beyond 50% if needed with 'expand'.
    */
   labelDisplay: PropTypes.oneOf(DISPLAYS),
+  /**
+   * Handler for click on a suggestItem.
+   */
+  onClick: PropTypes.func,
 };
 
 EuiSuggestItem.defaultProps = {